+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Generate session IDs with `uid-safe`, faster and even less collisions
+
 1.4.0 / 2014-06-17
 ==================